refactor icon component
