+++ src/components/Button/Button.jsx
@@ -1,11 +1,11 @@
 import Icon from "../Icon";
-import getPath, { getClassName } from "./utils";
+import { getClassName } from "./utils";
 import "./Button.css";
 
 const Button = (props) => {
     return <button className={getClassName(props.variant)}>
         {props.text}
-        <Icon path={getPath(props.icon)} />
+        <Icon icon={props.icon} />
     </button>
 }
